move style related functions to styles
