+++ cleopatra/array.py
@@ -11,7 +11,7 @@ from numpy_utils.filter import get_indices2
 # from matplotlib import gridspec
 from matplotlib.animation import FuncAnimation
 from matplotlib.ticker import LogFormatter
-
+from cleopatra.styles import Scale, MidpointNormalize
 
 DEFAULT_OPTIONS = dict(
     figsize=(8, 8),
@@ -121,7 +121,7 @@ class Array:
 
             array[mask] = np.nan
 
-        self.exclude_value = exclude_value
+        self._exclude_value = exclude_value
         self._vmin = np.nanmin(array)
         self._vmax = np.nanmax(array)
         self.arr = array
@@ -136,7 +136,7 @@ class Array:
             no_elem = np.size(array[:, :]) - np.count_nonzero((array[np.isnan(array)]))
 
         self.no_elem = no_elem
-        self.default_options = DEFAULT_OPTIONS
+        self._default_options = DEFAULT_OPTIONS
 
     @property
     def vmin(self):
@@ -147,6 +147,20 @@ class Array:
     def vmax(self):
         """max value in the array"""
         return self._vmax
+
+    @property
+    def exclude_value(self):
+        """exclude_value"""
+        return self._exclude_value
+
+    @exclude_value.setter
+    def exclude_value(self, value):
+        self._exclude_value = value
+
+    @property
+    def default_options(self):
+        """Default plot options"""
+        return self._default_options
 
     @staticmethod
     def get_line_style(style: Union[str, int] = "loosely dotted"):
@@ -973,154 +987,3 @@ class Array:
         # plt.tight_layout()
 
         return (ax1, ax2), fig
-
-
-class Scale:
-    """different scale object."""
-
-    def __init__(self):
-        """Different scale object."""
-        pass
-
-    @staticmethod
-    def log_scale(minval, maxval):
-        """log_scale.
-
-            logarithmic scale
-
-        Parameters
-        ----------
-        minval
-        maxval
-
-        Returns
-        -------
-        """
-
-        def scalar(val):
-            """scalar.
-
-                scalar
-
-            Parameters
-            ----------
-            val
-
-            Returns
-            -------
-            """
-            val = val + abs(minval) + 1
-            return np.log10(val)
-
-        return scalar
-
-    @staticmethod
-    def power_scale(minval, maxval):
-        """power_scale.
-
-            power scale
-
-        Parameters
-        ----------
-        minval
-        maxval
-
-        Returns
-        -------
-        """
-
-        def scalar(val):
-            val = val + abs(minval) + 1
-            return (val / 1000) ** 2
-
-        return scalar
-
-    @staticmethod
-    def identity_scale(minval, maxval):
-        """identity_scale.
-
-            identity_scale
-
-        Parameters
-        ----------
-        minval
-        maxval
-
-        Returns
-        -------
-        """
-
-        def scalar(val):
-            return 2
-
-        return scalar
-
-    @staticmethod
-    def rescale(OldValue, OldMin, OldMax, NewMin, NewMax):
-        """Rescale.
-
-        Rescale nethod rescales a value between two boundaries to a new value
-        bewteen two other boundaries
-        inputs:
-            1-OldValue:
-                [float] value need to transformed
-            2-OldMin:
-                [float] min old value
-            3-OldMax:
-                [float] max old value
-            4-NewMin:
-                [float] min new value
-            5-NewMax:
-                [float] max new value
-        output:
-            1-NewValue:
-                [float] transformed new value
-        """
-        OldRange = OldMax - OldMin
-        NewRange = NewMax - NewMin
-        NewValue = (((OldValue - OldMin) * NewRange) / OldRange) + NewMin
-
-        return NewValue
-
-
-class MidpointNormalize(colors.Normalize):
-    """MidpointNormalize.
-
-    !TODO needs docs
-    """
-
-    def __init__(self, vmin=None, vmax=None, midpoint=None, clip=False):
-        """MidpointNormalize.
-
-        Parameters
-        ----------
-        vmin
-        vmax
-        midpoint
-        clip
-        """
-        self.midpoint = midpoint
-        colors.Normalize.__init__(self, vmin, vmax, clip)
-
-    def __call__(self, value, clip=None):
-        """MidpointNormalize.
-
-        ! TODO needs docs
-
-        Parameters
-        ----------
-        value : TYPE
-            DESCRIPTION.
-        clip : TYPE, optional
-            DESCRIPTION. The default is None.
-
-        Returns
-        -------
-        TYPE
-            DESCRIPTION.
-        """
-        # I'm ignoring masked values and all kinds of edge cases to make a
-        # simple example...
-        x, y = [self.vmin, self.midpoint, self.vmax], [0, 0.5, 1]
-
-        return np.ma.masked_array(np.interp(value, x, y))
